requirements.txt
flask==2.0.1
flask-cors==3.0.10
tensorflow==2.9.0
tensorflow-hub==0.12.0
numpy==1.23.0
pillow==9.2.0